requirements.txt
wikipedia
pywhatkit
screen_brightness_control
pycaw
numpy
pandas
scipy
opencv-python
PyDictionary
pocketsphinx
SpeechRecognition
google
folium
ffmpeg
pywin32
openai
openweather
twilio
geolocation-python
mouse
keyboard
pyautogui
shazamio
newspaper3k
pathlib
aspose-words
whapi
schedule
pyjokes
spotipy
pygame